+++ acceptance-tests/src/test/java/tech/pegasys/signers/secp256k1/tests/multikey/signing/MultiKeyAzureTransactionSignerAcceptanceTest.java
@@ -27,8 +27,6 @@ public class MultiKeyAzureTransactionSignerAcceptanceTest
   static final String clientId = System.getenv("AZURE_CLIENT_ID");
   static final String clientSecret = System.getenv("AZURE_CLIENT_SECRET");
   static final String keyVaultName = System.getenv("AZURE_KEY_VAULT_NAME");
-  static final String FILENAME =
-      "09b02f8a5fddd222ade4ea4528faefc399623af3f736be3c44f03e2df22fb792f3931a4d9573d333ca74343305762a753388c3422a86d98b713fc91c1ea04842";
 
   @BeforeAll
   public static void checkAzureCredentials() {
@@ -40,7 +38,7 @@ public class MultiKeyAzureTransactionSignerAcceptanceTest
   @Test
   public void azureLoadedFromMultiKeyCanSign(@TempDir Path tomlDirectory) {
     createAzureTomlFileAt(
-        tomlDirectory.resolve("arbitrary_prefix" + FILENAME + ".toml"),
+        tomlDirectory.resolve(PUBLIC_KEY_HEX_STRING + ".toml"),
         clientId,
         clientSecret,
         keyVaultName);
