+++ acceptance-tests/src/testFixtures/java/tech.pegasys.signers.secp256k1/DefaultFileSelector.java
@@ -10,27 +10,26 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.signers.secp256k1.common;
+package tech.pegasys.signers.secp256k1;
 
+import tech.pegasys.signers.secp256k1.api.FileSelector;
 import tech.pegasys.signers.secp256k1.api.PublicKey;
 
-import org.apache.tuweni.bytes.Bytes;
+import java.nio.file.DirectoryStream.Filter;
+import java.nio.file.Path;
 
-public class StubbedPublicKey implements PublicKey {
+public class DefaultFileSelector implements FileSelector<PublicKey> {
 
-  private final String keyHexString;
-
-  public StubbedPublicKey(final String keyHexString) {
-    this.keyHexString = keyHexString;
+  @Override
+  public Filter<Path> getAllConfigFilesFilter() {
+    return entry -> entry.getFileName().toString().endsWith("toml");
   }
 
   @Override
-  public byte[] getValue() {
-    return Bytes.fromHexString(keyHexString).toArrayUnsafe();
-  }
-
-  @Override
-  public String toString() {
-    return keyHexString;
+  public Filter<Path> getSpecificConfigFileFilter(final PublicKey publicKey) {
+    return entry -> {
+      final String filename = publicKey.toString().substring(2); // remove 0x prefix
+      return entry.getFileName().toString().endsWith(filename + ".toml");
+    };
   }
 }
